requirements.txt
appnope==0.1.3
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
botocore==1.29.37
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.8.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
httplib2==0.21.0
huggingface-hub==0.11.1
idna==3.4
ipython==8.7.0
jedi==0.18.2
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.24.0
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycurl==7.45.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
s3transfer==0.6.0
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.0
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
